Refactor product-composable with sf-prefix and fix import paths
