--- packages/data/composables/useProduct/index.ts
@@ -1,2 +1,0 @@
-export * from './useProduct';
-export * from './types';
