+++ packages/data/composables/useSfProduct/index.ts
@@ -1,0 +1,2 @@
+export * from './useSfProduct';
+export * from './types';
