+++ packages/theme/components/ProductAccordion/__tests__/ProductAccordion.spec.ts
@@ -1,6 +1,6 @@
 import { mount } from '@vue/test-utils';
 import ProductAccordion from '../../ProductAccordion/ProductAccordion.vue';
-import { mockProduct } from '@crearis/data-main/composables/useSfProductAttribute/__tests__/useProduct.mock';
+import { mockProduct } from '@crearis/data-main/composables/useSfProductAttribute/__tests__/useSfProduct.mock';
 
 describe('<ProductAccordion />', () => {
   it('should render component', () => {
