+++ packages/theme/components/ProductProperties/__tests__/ProductProperties.spec.ts
@@ -1,6 +1,6 @@
 import { mount } from '@vue/test-utils';
 import ProductProperties from '../ProductProperties.vue';
-import { mockProduct } from '@crearis/data-main/composables/useSfProductAttribute/__tests__/useProduct.mock';
+import { mockProduct } from '@crearis/data-main/composables/useSfProductAttribute/__tests__/useSfProduct.mock';
 
 describe('<ProductProperties />', () => {
   it('should render component', () => {
